requirements.txt
graphene==2.1.3
elasticsearch==6.3.1
python-dateutil==2.8.0
Flask==1.0.2
Flask-GraphQL==2.0.0
